requirements.txt
python-binance
vaderSentiment
python-dotenv
requests